pyproject.toml
[tool.poetry]
name = "drf youtube"
version = "0.1.0"
description = ""
authors = ["Vasker <v92prog@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.13"
Pillow = "^9.1.0"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
django-ckeditor = "^6.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
